backend/requirements.txt
fastapi
uvicorn
anthropic
openai
httpx
beautifulsoup4
python-multipart
umap-learn
numpy
pandas
openpyxl
pymupdf
scikit-learn
pydantic
python-dotenv
youtube-transcript-api
xlrd
